requirements.txt
# MultiGridDet Requirements
# Core dependencies for MultiGridDet object detection

# Deep Learning Framework
tensorflow>=2.17.0,<3.0.0
keras>=3.0.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0
albumentations>=1.0.0

# Data Processing
numpy>=1.21.0
scipy>=1.7.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Utilities
tqdm>=4.60.0
pathlib2>=2.3.0; python_version < "3.4"

# Optional: For advanced augmentation
imgaug>=0.4.0

# Optional: For model optimization
# tensorflow-model-optimization>=0.7.0

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=6.0.0
# pytest-cov>=2.10.0
# black>=21.0.0
# flake8>=3.8.0





